feat: refactor teacher management pages to use context for data retrieval and improve edit functionality
